requirements.txt
astroid==2.3.3
grpcio==1.26.0
grpcio-tools==1.26.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
protobuf==3.11.2
pylint==2.4.4
six==1.13.0
typed-ast==1.4.0
wrapt==1.11.2
